backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic>=2.13,<3
requests==2.32.5
beautifulsoup4==4.13.4
sentence-transformers==5.1.0
faiss-cpu==1.12.0
numpy==2.3.2
python-dotenv==1.1.1
